moved sidebar components to colocate
