+++ apps/nextjs/src/app/(dashboard)/dashboard/_components/dashboard-content-skeleton.tsx
@@ -3,7 +3,7 @@ export default function DashboardContentSkeleton() {
     <section
       role="status"
       aria-label="Loading dashboard content"
-      className="flex flex-1 flex-col gap-4 p-4 pt-0"
+      className="flex flex-1 flex-col gap-4 p-4"
     >
       <div className="grid auto-rows-min gap-4 md:grid-cols-3">
         <div
