+++ apps/nextjs/src/app/(dashboard)/dashboard/_components/sidebar/nav-projects.tsx
