+++ apps/nextjs/src/app/(dashboard)/dashboard/page.tsx
@@ -1,7 +1,7 @@
+import DashboardContentSkeleton from "./_components/dashboard-content-skeleton";
+
 export default function DashboardProfiles() {
   return (
-    <div>
-      <h1>DASHBOARD PAGE</h1>
-    </div>
+    <DashboardContentSkeleton />
   );
 }
