+++ packages/ui/src/components/separator.tsx
@@ -21,7 +21,7 @@ const Separator = React.forwardRef<SeparatorRef, SeparatorProps>(
       orientation={orientation}
       className={cn(
         "shrink-0 bg-border",
-        orientation === "horizontal" ? "h-[1px] w-full" : "h-full w-[1px]",
+        orientation === "horizontal" ? "h-[1px] w-full" : "h-full w-[8px]",
         className,
       )}
       {...props}
